+++ CHANGELOG.md
@@ -54,7 +54,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   ([#1671](https://github.com/open-telemetry/opentelemetry-python/pull/1671))
 - Initial documentation for environment variables
   ([#1680](https://github.com/open-telemetry/opentelemetry-python/pull/1680))
-- Change Zipkin exporter to obtain service_name from span
+- Change Zipkin exporter to obtain service.name from span
   ([#1696](https://github.com/open-telemetry/opentelemetry-python/pull/1696))
 
 ### Removed